requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.36.2
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.0.1
dill==0.3.8
distro==1.9.0
filelock==3.16.1
fonttools==4.56.0
frozenlist==1.4.1
fsspec==2024.6.1
func_timeout==4.3.5
git-filter-repo==2.47.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
jiter==0.5.0
kiwisolver==1.4.8
matplotlib==3.10.0
multidict==6.1.0
multiprocess==0.70.16
numpy==2.1.1
openai==1.51.0
packaging==24.1
pandas==2.2.3
pillow==11.1.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
tokenizers==0.20.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wheel==0.44.0
xxhash==3.5.0
yarl==1.13.1